changed project structure-everything about relocated to package "tests" in the root dir
