--- fixtures/rest_api_fixt.py
@@ -1,3 +1,0 @@
-import pytest
-import test_data.endpoints
-from test_data.endpoints import gorest_users_url
