--- test_suite1/conftest.py
@@ -1,2 +1,0 @@
-from fixtures.drivers import *
-from fixtures.pages import *
