--- test_suite2/conftest.py
@@ -1,1 +1,0 @@
-from fixtures.rest_api_fixt import *
