+++ tests/fixtures/drivers.py
@@ -2,8 +2,8 @@ import pytest
 from time import sleep
 from webdriver_manager.chrome import ChromeDriverManager
 from selenium import webdriver
-import test_data.endpoints
-from pages.pageobjects import FavoritePage
+import tests.test_data.endpoints
+from tests.pages.pageobjects import FavoritePage
 
 
 @pytest.fixture
@@ -13,7 +13,7 @@ def driver(request):
     driver.implicitly_wait(10)
 
     def fin():
-        driver.get(test_data.endpoints.favorites_url)
+        driver.get(tests.test_data.endpoints.favorites_url)
         # очищается список избранных после каждого теста
         for cross_button_index in FavoritePage(driver).del_favs_button:
             sleep(1)  # пока что ничего лучшего тут не вышло(
