+++ tests/fixtures/pages.py
@@ -1,13 +1,13 @@
 import pytest
-from pages.pageobjects import MainPage
-from fixtures.drivers import driver
-import test_data.endpoints
-import common
+from tests.pages.pageobjects import MainPage
+from tests.fixtures.drivers import driver
+import tests.test_data.endpoints
+from tests import common
 
 
 @pytest.fixture
 def logined_page(driver: driver):
-    driver.get(test_data.endpoints.velosipednye_shiny_url)
+    driver.get(tests.test_data.endpoints.velosipednye_shiny_url)
     main_page = MainPage(driver)
     main_page.comein_link.click()
     common.login_steps(main_page)
@@ -16,5 +16,5 @@ def logined_page(driver: driver):
 
 @pytest.fixture
 def main_page(driver: driver):
-    driver.get(test_data.endpoints.velosipednye_shiny_url)
+    driver.get(tests.test_data.endpoints.velosipednye_shiny_url)
     return MainPage(driver)
