+++ tests/test_suite1/conftest.py
@@ -1,0 +1,2 @@
+from tests.fixtures.drivers import *
+from tests.fixtures.pages import *
